Add files via upload
Used pandas, numpy, and matplotlib libraries. For the unit projects, the goals were to create visualizations for the data we were given and as for the unit challenge the goal was to create a function and have it state whether or not a word was an isogram using code we had learned.